refactor: rename `charts` folder for consistency
