+++ packages/picasso.js/src/core/chart/formatter/index.js
